# Use efficient ARM fp16 dot product for gemm_transa_ general case